Update Chapter1Practice.js
Want to build out a function that dioes it all in one line basically. I am not going to put serious sustained effort into this because of how It does not do much in terms of scaling or purpose of program, But I will Throw in additional lines of code when I learn more. Maybe one day even write this into one line of code itself.